refactor: split dce guards and writes
