+++ src/optimize/control/dce.rs
@@ -1,10 +1,14 @@
 use super::*;
 
+mod guards;
 mod methods;
 mod state;
+mod writes;
 
 pub(crate) use methods::{dce_method, dce_method_without_context};
-use state::{ConditionGuard, ExactGuard, GuardLiteral, GuardState, TailSinkTarget};
+use guards::*;
+use state::{GuardLiteral, GuardState, TailSinkTarget};
+use writes::*;
 
 pub(crate) fn dce_block(body: Vec<Stmt>) -> Vec<Stmt> {
     dce_block_with_guards(body, GuardState::default())
@@ -1457,1055 +1461,5 @@ fn dce_stmt_with_guards(stmt: Stmt, guards: &GuardState) -> Vec<Stmt> {
             span,
         }],
         kind => vec![Stmt { kind, span }],
-    }
-}
-
-fn guard_variable_name(condition: &Expr) -> Option<(&str, bool)> {
-    match &condition.kind {
-        ExprKind::Variable(name) => Some((name.as_str(), true)),
-        ExprKind::Not(inner) => match &inner.kind {
-            ExprKind::Variable(name) => Some((name.as_str(), false)),
-            _ => None,
-        },
-        _ => None,
-    }
-}
-
-fn scalar_guard_value(expr: &Expr) -> Option<GuardLiteral> {
-    match &expr.kind {
-        ExprKind::BoolLiteral(value) => Some(GuardLiteral::Bool(*value)),
-        ExprKind::Null => Some(GuardLiteral::Null),
-        ExprKind::IntLiteral(value) => Some(GuardLiteral::Int(*value)),
-        ExprKind::FloatLiteral(value) => Some(GuardLiteral::Float(value.to_bits())),
-        ExprKind::StringLiteral(value) => Some(GuardLiteral::String(value.clone())),
-        _ => None,
-    }
-}
-
-fn strict_scalar_guard(condition: &Expr) -> Option<(&str, GuardLiteral, bool)> {
-    let ExprKind::BinaryOp { left, op, right } = &condition.kind else {
-        return None;
-    };
-
-    let (name, value) = match (&left.kind, &right.kind) {
-        (ExprKind::Variable(name), _) => (name.as_str(), scalar_guard_value(right)?),
-        (_, ExprKind::Variable(name)) => (name.as_str(), scalar_guard_value(left)?),
-        _ => return None,
-    };
-
-    match op {
-        BinOp::StrictEq => Some((name, value, true)),
-        BinOp::StrictNotEq => Some((name, value, false)),
-        _ => None,
-    }
-}
-
-fn guard_literal_truthy(value: &GuardLiteral) -> bool {
-    match value {
-        GuardLiteral::Bool(value) => *value,
-        GuardLiteral::Null => false,
-        GuardLiteral::Int(value) => *value != 0,
-        GuardLiteral::Float(bits) => f64::from_bits(*bits) != 0.0,
-        GuardLiteral::String(value) => !value.is_empty() && value != "0",
-    }
-}
-
-fn known_exact_guard<'a>(guards: &'a GuardState, name: &str) -> Option<&'a GuardLiteral> {
-    guards
-        .exact_guards
-        .iter()
-        .find(|known| known.name == name)
-        .map(|known| &known.value)
-}
-
-fn has_excluded_guard(guards: &GuardState, name: &str, value: &GuardLiteral) -> bool {
-    guards
-        .excluded_guards
-        .iter()
-        .any(|known| known.name == name && known.value == *value)
-}
-
-fn known_condition_value(condition: &Expr, guards: &GuardState) -> Option<bool> {
-    let mut visiting = Vec::new();
-    known_condition_value_inner(condition, guards, &mut visiting)
-}
-
-fn known_condition_value_inner(
-    condition: &Expr,
-    guards: &GuardState,
-    visiting: &mut Vec<Expr>,
-) -> Option<bool> {
-    if visiting.iter().any(|known| known == condition) {
-        return None;
-    }
-
-    visiting.push(condition.clone());
-    let value = if let Some(value) = known_condition_value_base(condition, guards, visiting) {
-        Some(value)
-    } else {
-        infer_condition_value_from_composite_guards(condition, guards, visiting)
-    };
-    visiting.pop();
-
-    value
-}
-
-fn known_condition_value_base(
-    condition: &Expr,
-    guards: &GuardState,
-    visiting: &mut Vec<Expr>,
-) -> Option<bool> {
-    if let Some(value) = guards
-        .condition_guards
-        .iter()
-        .find(|known| known.condition == *condition)
-        .map(|known| known.value)
-    {
-        return Some(value);
-    }
-
-    if let ExprKind::Not(inner) = &condition.kind {
-        return known_condition_value_inner(inner, guards, visiting).map(|value| !value);
-    }
-
-    if let ExprKind::BinaryOp { left, op, right } = &condition.kind {
-        match op {
-            BinOp::And => match (
-                known_condition_value_inner(left, guards, visiting),
-                known_condition_value_inner(right, guards, visiting),
-            ) {
-                (Some(false), _) | (_, Some(false)) => return Some(false),
-                (Some(true), Some(true)) => return Some(true),
-                _ => {}
-            },
-            BinOp::Or => match (
-                known_condition_value_inner(left, guards, visiting),
-                known_condition_value_inner(right, guards, visiting),
-            ) {
-                (Some(true), _) | (_, Some(true)) => return Some(true),
-                (Some(false), Some(false)) => return Some(false),
-                _ => {}
-            },
-            _ => {}
-        }
-    }
-
-    if let Some((name, truthy_if_true)) = guard_variable_name(condition) {
-        if let Some(value) = known_exact_guard(guards, name) {
-            return Some(guard_literal_truthy(value) == truthy_if_true);
-        }
-        if guards.bool_true_vars.iter().any(|known| known == name)
-            || guards.truthy_vars.iter().any(|known| known == name)
-        {
-            return Some(truthy_if_true);
-        }
-        if guards.bool_false_vars.iter().any(|known| known == name)
-            || guards.falsy_vars.iter().any(|known| known == name)
-        {
-            return Some(!truthy_if_true);
-        }
-    }
-
-    if let Some((name, compared_value, expects_equal)) = strict_scalar_guard(condition) {
-        if let Some(known) = known_exact_guard(guards, name) {
-            return Some((known == &compared_value) == expects_equal);
-        }
-        if has_excluded_guard(guards, name, &compared_value) {
-            return Some(!expects_equal);
-        }
-    }
-
-    None
-}
-
-fn expr_contains_subexpr(expr: &Expr, target: &Expr) -> bool {
-    if expr == target {
-        return true;
-    }
-
-    match &expr.kind {
-        ExprKind::Not(inner) | ExprKind::Negate(inner) | ExprKind::BitNot(inner) => {
-            expr_contains_subexpr(inner, target)
-        }
-        ExprKind::BinaryOp { left, right, .. } => {
-            expr_contains_subexpr(left, target) || expr_contains_subexpr(right, target)
-        }
-        _ => false,
-    }
-}
-
-fn infer_child_value_from_composite_guard(
-    op: BinOp,
-    composite_value: bool,
-    sibling_value: Option<bool>,
-) -> Option<bool> {
-    match (op, composite_value, sibling_value) {
-        (BinOp::And, true, _) => Some(true),
-        (BinOp::Or, false, _) => Some(false),
-        (BinOp::And, false, Some(true)) => Some(false),
-        (BinOp::Or, true, Some(false)) => Some(true),
-        _ => None,
-    }
-}
-
-fn infer_condition_value_from_composite_tree(
-    condition: &Expr,
-    composite: &Expr,
-    composite_value: bool,
-    guards: &GuardState,
-    visiting: &mut Vec<Expr>,
-) -> Option<bool> {
-    if composite == condition {
-        return Some(composite_value);
-    }
-
-    if let ExprKind::Not(inner) = &composite.kind {
-        return infer_condition_value_from_composite_tree(
-            condition,
-            inner,
-            !composite_value,
-            guards,
-            visiting,
-        );
-    }
-
-    let ExprKind::BinaryOp { left, op, right } = &composite.kind else {
-        return None;
-    };
-
-    let left_contains = expr_contains_subexpr(left, condition);
-    let right_contains = expr_contains_subexpr(right, condition);
-    let (candidate, sibling) = match (left_contains, right_contains) {
-        (true, false) => (&**left, &**right),
-        (false, true) => (&**right, &**left),
-        _ => return None,
-    };
-
-    let candidate_value =
-        infer_child_value_from_composite_guard(
-            op.clone(),
-            composite_value,
-            known_condition_value_inner(sibling, guards, visiting),
-        )?;
-
-    infer_condition_value_from_composite_tree(
-        condition,
-        candidate,
-        candidate_value,
-        guards,
-        visiting,
-    )
-}
-
-fn infer_condition_value_from_composite_guards(
-    condition: &Expr,
-    guards: &GuardState,
-    visiting: &mut Vec<Expr>,
-) -> Option<bool> {
-    for known in &guards.condition_guards {
-        if !expr_contains_subexpr(&known.condition, condition) {
-            continue;
-        }
-
-        if let Some(value) = infer_condition_value_from_composite_tree(
-            condition,
-            &known.condition,
-            known.value,
-            guards,
-            visiting,
-        ) {
-            return Some(value);
-        }
-    }
-
-    None
-}
-
-fn clear_guards_for_name(guards: &mut GuardState, name: &str) {
-    guards.truthy_vars.retain(|known| known != name);
-    guards.falsy_vars.retain(|known| known != name);
-    guards.bool_true_vars.retain(|known| known != name);
-    guards.bool_false_vars.retain(|known| known != name);
-    guards.exact_guards.retain(|known| known.name != name);
-    guards.excluded_guards.retain(|known| known.name != name);
-    guards
-        .condition_guards
-        .retain(|known| !known.names.iter().any(|known_name| known_name == name));
-}
-
-fn push_guard_name(names: &mut Vec<String>, name: &str) {
-    if !names.iter().any(|known| known == name) {
-        names.push(name.to_string());
-    }
-}
-
-fn record_truthy_guard(guards: &mut GuardState, name: &str, known_truthy: bool) {
-    guards.truthy_vars.retain(|known| known != name);
-    guards.falsy_vars.retain(|known| known != name);
-    if known_truthy {
-        push_guard_name(&mut guards.truthy_vars, name);
-    } else {
-        push_guard_name(&mut guards.falsy_vars, name);
-    }
-}
-
-fn record_exact_literal_guard(guards: &mut GuardState, name: &str, value: GuardLiteral) {
-    clear_guards_for_name(guards, name);
-    if let GuardLiteral::Bool(value) = &value {
-        if *value {
-            push_guard_name(&mut guards.bool_true_vars, name);
-        } else {
-            push_guard_name(&mut guards.bool_false_vars, name);
-        }
-    }
-    guards.exact_guards.push(ExactGuard {
-        name: name.to_string(),
-        value: value.clone(),
-    });
-    record_truthy_guard(guards, name, guard_literal_truthy(&value));
-}
-
-fn exact_literal_from_guard_branch(condition: &Expr, branch_taken: bool) -> Option<(&str, GuardLiteral)> {
-    let (name, compared_value, expects_equal) = strict_scalar_guard(condition)?;
-    match (expects_equal, branch_taken) {
-        (true, true) => Some((name, compared_value)),
-        (false, false) => Some((name, compared_value)),
-        _ => None,
-    }
-}
-
-fn excluded_literal_from_guard_branch(
-    condition: &Expr,
-    branch_taken: bool,
-) -> Option<(&str, GuardLiteral)> {
-    let (name, compared_value, expects_equal) = strict_scalar_guard(condition)?;
-    match (expects_equal, branch_taken) {
-        (true, false) => Some((name, compared_value)),
-        (false, true) => Some((name, compared_value)),
-        _ => None,
-    }
-}
-
-fn record_excluded_literal_guard(guards: &mut GuardState, name: &str, value: GuardLiteral) {
-    if !guards
-        .excluded_guards
-        .iter()
-        .any(|known| known.name == name && known.value == value)
-    {
-        guards.excluded_guards.push(ExactGuard {
-            name: name.to_string(),
-            value,
-        });
-    }
-}
-
-fn collect_trackable_condition_names(expr: &Expr, names: &mut Vec<String>) -> bool {
-    match &expr.kind {
-        ExprKind::Variable(name) => {
-            push_guard_name(names, name);
-            true
-        }
-        ExprKind::BoolLiteral(_)
-        | ExprKind::Null
-        | ExprKind::IntLiteral(_)
-        | ExprKind::FloatLiteral(_)
-        | ExprKind::StringLiteral(_) => true,
-        ExprKind::Not(inner) | ExprKind::Negate(inner) | ExprKind::BitNot(inner) => {
-            collect_trackable_condition_names(inner, names)
-        }
-        ExprKind::BinaryOp { left, right, .. } => {
-            collect_trackable_condition_names(left, names)
-                && collect_trackable_condition_names(right, names)
-        }
-        _ => false,
-    }
-}
-
-fn inverse_comparison_op(op: &BinOp) -> Option<BinOp> {
-    match op {
-        BinOp::Eq => Some(BinOp::NotEq),
-        BinOp::NotEq => Some(BinOp::Eq),
-        BinOp::StrictEq => Some(BinOp::StrictNotEq),
-        BinOp::StrictNotEq => Some(BinOp::StrictEq),
-        BinOp::Lt => Some(BinOp::GtEq),
-        BinOp::Gt => Some(BinOp::LtEq),
-        BinOp::LtEq => Some(BinOp::Gt),
-        BinOp::GtEq => Some(BinOp::Lt),
-        _ => None,
-    }
-}
-
-fn comparison_inverse_is_total(op: &BinOp) -> bool {
-    matches!(op, BinOp::Eq | BinOp::NotEq | BinOp::StrictEq | BinOp::StrictNotEq)
-}
-
-fn condition_guard_forms(condition: &Expr, value: bool) -> Vec<(Expr, bool)> {
-    let mut forms = Vec::new();
-
-    match &condition.kind {
-        ExprKind::Not(inner) => {
-            if let ExprKind::BinaryOp { left, op, right } = &inner.kind {
-                let de_morgan_op = match op {
-                    BinOp::And => Some(BinOp::Or),
-                    BinOp::Or => Some(BinOp::And),
-                    _ => None,
-                };
-
-                if let Some(de_morgan_op) = de_morgan_op {
-                    forms.push((
-                        Expr::binop(
-                            invert_condition((**left).clone()),
-                            de_morgan_op,
-                            invert_condition((**right).clone()),
-                        ),
-                        value,
-                    ));
-                }
-            }
-        }
-        ExprKind::BinaryOp { left, op, right } => {
-            if let Some(inverse_op) = inverse_comparison_op(op) {
-                if value || comparison_inverse_is_total(op) {
-                    forms.push((
-                        Expr::binop((**left).clone(), inverse_op, (**right).clone()),
-                        !value,
-                    ));
-                }
-            }
-
-            let de_morgan_op = match op {
-                BinOp::And => Some(BinOp::Or),
-                BinOp::Or => Some(BinOp::And),
-                _ => None,
-            };
-
-            if let (
-                Some(de_morgan_op),
-                ExprKind::Not(left_inner),
-                ExprKind::Not(right_inner),
-            ) = (de_morgan_op, &left.kind, &right.kind)
-            {
-                forms.push((
-                    invert_condition(Expr::binop(
-                        (**left_inner).clone(),
-                        de_morgan_op,
-                        (**right_inner).clone(),
-                    )),
-                    value,
-                ));
-            }
-        }
-        _ => {}
-    }
-
-    forms
-}
-
-fn upsert_condition_guard(
-    guards: &mut GuardState,
-    condition: Expr,
-    value: bool,
-    names: &[String],
-) {
-    if let Some(existing) = guards
-        .condition_guards
-        .iter_mut()
-        .find(|known| known.condition == condition)
-    {
-        existing.value = value;
-        existing.names = names.to_vec();
-        return;
-    }
-
-    guards.condition_guards.push(ConditionGuard {
-        condition,
-        value,
-        names: names.to_vec(),
-    });
-}
-
-fn record_condition_guard(guards: &mut GuardState, condition: &Expr, value: bool) {
-    let effect = expr_effect(condition);
-    if effect.has_side_effects || effect.may_throw {
-        return;
-    }
-
-    let mut names = Vec::new();
-    if !collect_trackable_condition_names(condition, &mut names) {
-        return;
-    }
-
-    upsert_condition_guard(guards, condition.clone(), value, &names);
-    for (equivalent, equivalent_value) in condition_guard_forms(condition, value) {
-        let equivalent_effect = expr_effect(&equivalent);
-        if equivalent_effect.has_side_effects || equivalent_effect.may_throw {
-            continue;
-        }
-        upsert_condition_guard(guards, equivalent, equivalent_value, &names);
-    }
-}
-
-fn extend_guards_for_switch_case(subject: &Expr, patterns: &[Expr], guards: &GuardState) -> GuardState {
-    let [pattern] = patterns else {
-        return guards.clone();
-    };
-
-    match &subject.kind {
-        ExprKind::BoolLiteral(subject_bool) => extend_guards(guards, pattern, *subject_bool),
-        ExprKind::Variable(name) => {
-            let mut next = guards.clone();
-            if let ExprKind::BoolLiteral(pattern_bool) = pattern.kind {
-                record_truthy_guard(&mut next, name, pattern_bool);
-            }
-            next
-        }
-        _ => guards.clone(),
-    }
-}
-
-fn extend_guards_for_switch_case_no_match(
-    subject_value: &ScalarValue,
-    patterns: &[Expr],
-    guards: &GuardState,
-) -> GuardState {
-    let ScalarValue::Bool(subject_bool) = subject_value else {
-        return guards.clone();
-    };
-
-    patterns.iter().fold(guards.clone(), |guards, pattern| {
-        extend_guards(&guards, pattern, !subject_bool)
-    })
-}
-
-fn extend_guards_for_switch_case_no_match_subject(
-    subject: &Expr,
-    patterns: &[Expr],
-    guards: &GuardState,
-) -> GuardState {
-    if let Some(subject_value) = known_scalar_subject_value(subject, guards) {
-        if matches!(subject_value, ScalarValue::Bool(_)) {
-            return extend_guards_for_switch_case_no_match(&subject_value, patterns, guards);
-        }
-    }
-
-    let ExprKind::Variable(name) = &subject.kind else {
-        return guards.clone();
-    };
-
-    patterns.iter().fold(guards.clone(), |mut guards, pattern| {
-        match &pattern.kind {
-            ExprKind::BoolLiteral(pattern_bool) => {
-                record_truthy_guard(&mut guards, name, !pattern_bool);
-            }
-            _ => {
-                if let Some(pattern_value) = scalar_guard_value(pattern) {
-                    record_excluded_literal_guard(&mut guards, name, pattern_value);
-                }
-            }
-        }
-        guards
-    })
-}
-
-fn extend_guards(guards: &GuardState, condition: &Expr, branch_taken: bool) -> GuardState {
-    let mut next = if let ExprKind::Not(inner) = &condition.kind {
-        extend_guards(guards, inner, !branch_taken)
-    } else if let ExprKind::BinaryOp { left, op, right } = &condition.kind {
-        match (op, branch_taken) {
-            (BinOp::And, true) => {
-                let left_true = extend_guards(guards, left, true);
-                extend_guards(&left_true, right, true)
-            }
-            (BinOp::And, false) => {
-                if matches!(known_condition_value(left, guards), Some(true)) {
-                    let left_true = extend_guards(guards, left, true);
-                    extend_guards(&left_true, right, false)
-                } else {
-                    guards.clone()
-                }
-            }
-            (BinOp::Or, false) => {
-                let left_false = extend_guards(guards, left, false);
-                extend_guards(&left_false, right, false)
-            }
-            (BinOp::Or, true) => {
-                if matches!(known_condition_value(left, guards), Some(false)) {
-                    let left_false = extend_guards(guards, left, false);
-                    extend_guards(&left_false, right, true)
-                } else {
-                    guards.clone()
-                }
-            }
-            _ => guards.clone(),
-        }
-    } else {
-        guards.clone()
-    };
-
-    if let Some((name, exact_value)) = exact_literal_from_guard_branch(condition, branch_taken) {
-        record_exact_literal_guard(&mut next, name, exact_value);
-    }
-
-    if let Some((name, excluded_value)) = excluded_literal_from_guard_branch(condition, branch_taken) {
-        record_excluded_literal_guard(&mut next, name, excluded_value);
-    }
-
-    record_condition_guard(&mut next, condition, branch_taken);
-
-    if let Some((name, truthy_if_true)) = guard_variable_name(condition) {
-        let known_truthy = if branch_taken { truthy_if_true } else { !truthy_if_true };
-        record_truthy_guard(&mut next, name, known_truthy);
-    };
-
-    next
-}
-
-fn invalidate_guards_for_stmt(stmt: &Stmt, guards: &mut GuardState) {
-    let mut written = Vec::new();
-    collect_written_names(stmt, &mut written);
-    if written.is_empty() {
-        return;
-    }
-
-    invalidate_guards_for_written_names(guards, &written);
-}
-
-fn invalidated_guards_for_block(guards: &GuardState, stmts: &[Stmt]) -> GuardState {
-    let mut written = Vec::new();
-    collect_written_names_in_block(stmts, &mut written);
-    if written.is_empty() {
-        return guards.clone();
-    }
-
-    let mut next = guards.clone();
-    invalidate_guards_for_written_names(&mut next, &written);
-    next
-}
-
-fn invalidated_guards_for_throw_paths(guards: &GuardState, stmts: &[Stmt]) -> GuardState {
-    if !block_may_throw(stmts) {
-        return guards.clone();
-    }
-
-    let mut written = Vec::new();
-    collect_written_names_on_throw_paths_in_block(stmts, vec![Vec::new()], &mut written, guards);
-    if written.is_empty() {
-        return guards.clone();
-    }
-
-    let mut next = guards.clone();
-    invalidate_guards_for_written_names(&mut next, &written);
-    next
-}
-
-fn invalidated_guards_for_finally_paths(
-    guards: &GuardState,
-    try_body: &[Stmt],
-    catches: &[crate::parser::ast::CatchClause],
-) -> GuardState {
-    let mut next = invalidated_guards_for_block(guards, try_body);
-    for catch in catches {
-        next = invalidated_guards_for_block(&next, &catch.body);
-        if let Some(variable) = catch.variable.as_deref() {
-            clear_guards_for_name(&mut next, variable);
-        }
-    }
-    next
-}
-
-fn collect_written_names_on_throw_paths_in_block(
-    stmts: &[Stmt],
-    mut incoming_paths: Vec<Vec<String>>,
-    written: &mut Vec<String>,
-    guards: &GuardState,
-) -> Vec<Vec<String>> {
-    let mut current_guards = guards.clone();
-    for stmt in stmts {
-        if incoming_paths.is_empty() {
-            break;
-        }
-
-        let mut next_paths = Vec::new();
-        for path in incoming_paths {
-            collect_written_names_on_throw_paths_in_stmt(
-                stmt,
-                path,
-                written,
-                &mut next_paths,
-                &current_guards,
-            );
-        }
-        incoming_paths = next_paths;
-        invalidate_guards_for_stmt(stmt, &mut current_guards);
-    }
-
-    incoming_paths
-}
-
-fn collect_written_names_on_throw_paths_in_stmt(
-    stmt: &Stmt,
-    path: Vec<String>,
-    written: &mut Vec<String>,
-    next_paths: &mut Vec<Vec<String>>,
-    guards: &GuardState,
-) {
-    match &stmt.kind {
-        StmtKind::If {
-            condition,
-            then_body,
-            elseif_clauses,
-            else_body,
-        } => {
-            if expr_effect(condition).may_throw {
-                merge_written_path(written, &path);
-            }
-            next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                then_body,
-                vec![path.clone()],
-                written,
-                &extend_guards(guards, condition, true),
-            ));
-            next_paths.extend(collect_written_names_on_throw_paths_in_if_false_path(
-                elseif_clauses,
-                else_body,
-                path,
-                written,
-                &extend_guards(guards, condition, false),
-            ));
-            return;
-        }
-        StmtKind::IfDef {
-            then_body,
-            else_body,
-            ..
-        } => {
-            next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                then_body,
-                vec![path.clone()],
-                written,
-                guards,
-            ));
-            if let Some(body) = else_body {
-                next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                    body,
-                    vec![path],
-                    written,
-                    guards,
-                ));
-            } else {
-                next_paths.push(path);
-            }
-            return;
-        }
-        StmtKind::Switch {
-            subject,
-            cases,
-            default,
-        } => {
-            let (direct_case_entries, direct_default_entry) =
-                direct_switch_entry_blocks(subject, cases, default.is_some(), guards);
-            let cfg = build_switch_cfg(cases, default);
-            let mut entry_blocks = direct_case_entries.clone();
-            if direct_default_entry {
-                if let Some(default_entry) = cfg.default_entry {
-                    entry_blocks.push(default_entry);
-                }
-            }
-            let reachable = collect_reachable_cfg_blocks(&cfg.blocks, &entry_blocks);
-
-            if expr_effect(subject).may_throw {
-                merge_written_path(written, &path);
-            }
-
-            let mut fallthrough_paths = Vec::new();
-            for (index, (patterns, body)) in cases.iter().enumerate() {
-                let direct_entry = direct_case_entries.contains(&index);
-                if !reachable.get(index).copied().unwrap_or_default() {
-                    fallthrough_paths.clear();
-                    continue;
-                }
-
-                if direct_entry {
-                    for pattern in patterns {
-                        if expr_effect(pattern).may_throw {
-                            merge_written_path(written, &path);
-                        }
-                    }
-                }
-
-                let mut incoming = Vec::new();
-                if direct_entry {
-                    incoming.push(path.clone());
-                }
-                incoming.extend(fallthrough_paths);
-                if incoming.is_empty() {
-                    fallthrough_paths = Vec::new();
-                } else {
-                    fallthrough_paths = collect_written_names_on_throw_paths_in_block(
-                        body,
-                        incoming,
-                        written,
-                        guards,
-                    );
-                }
-            }
-
-            if let Some(body) = default {
-                let default_entry = cfg.default_entry.unwrap();
-                if reachable.get(default_entry).copied().unwrap_or_default() {
-                    let mut incoming = Vec::new();
-                    if direct_default_entry {
-                        incoming.push(path.clone());
-                    }
-                    incoming.extend(fallthrough_paths);
-                    if !incoming.is_empty() {
-                        let _ = collect_written_names_on_throw_paths_in_block(
-                            body,
-                            incoming,
-                            written,
-                            guards,
-                        );
-                    }
-                }
-            }
-
-            if matches!(stmt_terminal_effect(stmt), TerminalEffect::FallsThrough) {
-                let mut fallthrough = path;
-                collect_written_names(stmt, &mut fallthrough);
-                next_paths.push(fallthrough);
-            }
-            return;
-        }
-        StmtKind::Try {
-            try_body,
-            catches,
-            finally_body,
-        } => {
-            next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                try_body,
-                vec![path.clone()],
-                written,
-                guards,
-            ));
-            for catch in catches {
-                let mut catch_path = path.clone();
-                if let Some(variable) = catch.variable.as_deref() {
-                    push_written_name(&mut catch_path, variable);
-                }
-                next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                    &catch.body,
-                    vec![catch_path],
-                    written,
-                    guards,
-                ));
-            }
-            if let Some(body) = finally_body {
-                next_paths.extend(collect_written_names_on_throw_paths_in_block(
-                    body,
-                    vec![path],
-                    written,
-                    guards,
-                ));
-            } else if matches!(stmt_terminal_effect(stmt), TerminalEffect::FallsThrough) {
-                let mut fallthrough = path;
-                collect_written_names(stmt, &mut fallthrough);
-                next_paths.push(fallthrough);
-            }
-            return;
-        }
-        _ => {}
-    }
-
-    if stmt_may_throw(stmt) {
-        merge_written_path(written, &path);
-    }
-
-    if matches!(stmt_terminal_effect(stmt), TerminalEffect::FallsThrough) {
-        let mut fallthrough = path;
-        collect_written_names(stmt, &mut fallthrough);
-        next_paths.push(fallthrough);
-    }
-}
-
-fn collect_written_names_on_throw_paths_in_if_false_path(
-    elseif_clauses: &[(Expr, Vec<Stmt>)],
-    else_body: &Option<Vec<Stmt>>,
-    path: Vec<String>,
-    written: &mut Vec<String>,
-    guards: &GuardState,
-) -> Vec<Vec<String>> {
-    let Some((condition, body)) = elseif_clauses.first() else {
-        return else_body
-            .as_ref()
-            .map(|body| {
-                collect_written_names_on_throw_paths_in_block(body, vec![path.clone()], written, guards)
-            })
-            .unwrap_or_else(|| vec![path]);
-    };
-
-    if expr_effect(condition).may_throw {
-        merge_written_path(written, &path);
-    }
-
-    let mut next_paths = collect_written_names_on_throw_paths_in_block(
-        body,
-        vec![path.clone()],
-        written,
-        &extend_guards(guards, condition, true),
-    );
-    next_paths.extend(collect_written_names_on_throw_paths_in_if_false_path(
-        &elseif_clauses[1..],
-        else_body,
-        path,
-        written,
-        &extend_guards(guards, condition, false),
-    ));
-    next_paths
-}
-
-fn merge_written_path(written: &mut Vec<String>, path: &[String]) {
-    for name in path {
-        push_written_name(written, name);
-    }
-}
-
-fn invalidate_guards_for_written_names(guards: &mut GuardState, written: &[String]) {
-    guards
-        .truthy_vars
-        .retain(|name| !written.iter().any(|written_name| written_name == name));
-    guards
-        .falsy_vars
-        .retain(|name| !written.iter().any(|written_name| written_name == name));
-    guards
-        .bool_true_vars
-        .retain(|name| !written.iter().any(|written_name| written_name == name));
-    guards
-        .bool_false_vars
-        .retain(|name| !written.iter().any(|written_name| written_name == name));
-    guards
-        .exact_guards
-        .retain(|known| !written.iter().any(|written_name| written_name == &known.name));
-    guards
-        .excluded_guards
-        .retain(|known| !written.iter().any(|written_name| written_name == &known.name));
-    guards
-        .condition_guards
-        .retain(|known| !known.names.iter().any(|name| written.iter().any(|written_name| written_name == name)));
-}
-
-fn collect_written_names(stmt: &Stmt, written: &mut Vec<String>) {
-    match &stmt.kind {
-        StmtKind::Assign { name, .. }
-        | StmtKind::TypedAssign { name, .. }
-        | StmtKind::StaticVar { name, .. } => push_written_name(written, name),
-        StmtKind::ArrayAssign { array, .. } | StmtKind::ArrayPush { array, .. } => {
-            push_written_name(written, array)
-        }
-        StmtKind::ListUnpack { vars, .. } => {
-            for name in vars {
-                push_written_name(written, name);
-            }
-        }
-        StmtKind::ExprStmt(expr) => collect_expr_written_names(expr, written),
-        StmtKind::If {
-            then_body,
-            elseif_clauses,
-            else_body,
-            ..
-        } => {
-            collect_written_names_in_block(then_body, written);
-            for (_, body) in elseif_clauses {
-                collect_written_names_in_block(body, written);
-            }
-            if let Some(body) = else_body {
-                collect_written_names_in_block(body, written);
-            }
-        }
-        StmtKind::IfDef {
-            then_body,
-            else_body,
-            ..
-        } => {
-            collect_written_names_in_block(then_body, written);
-            if let Some(body) = else_body {
-                collect_written_names_in_block(body, written);
-            }
-        }
-        StmtKind::While { body, .. }
-        | StmtKind::DoWhile { body, .. }
-        | StmtKind::NamespaceBlock { body, .. } => collect_written_names_in_block(body, written),
-        StmtKind::For {
-            init,
-            update,
-            body,
-            ..
-        } => {
-            if let Some(stmt) = init {
-                collect_written_names(stmt, written);
-            }
-            if let Some(stmt) = update {
-                collect_written_names(stmt, written);
-            }
-            collect_written_names_in_block(body, written);
-        }
-        StmtKind::Foreach {
-            key_var,
-            value_var,
-            body,
-            ..
-        } => {
-            if let Some(name) = key_var {
-                push_written_name(written, name);
-            }
-            push_written_name(written, value_var);
-            collect_written_names_in_block(body, written);
-        }
-        StmtKind::Switch { cases, default, .. } => {
-            for (_, body) in cases {
-                collect_written_names_in_block(body, written);
-            }
-            if let Some(body) = default {
-                collect_written_names_in_block(body, written);
-            }
-        }
-        StmtKind::Try {
-            try_body,
-            catches,
-            finally_body,
-        } => {
-            collect_written_names_in_block(try_body, written);
-            for catch in catches {
-                if let Some(name) = &catch.variable {
-                    push_written_name(written, name);
-                }
-                collect_written_names_in_block(&catch.body, written);
-            }
-            if let Some(body) = finally_body {
-                collect_written_names_in_block(body, written);
-            }
-        }
-        _ => {}
-    }
-}
-
-fn collect_written_names_in_block(stmts: &[Stmt], written: &mut Vec<String>) {
-    for stmt in stmts {
-        collect_written_names(stmt, written);
-    }
-}
-
-fn collect_expr_written_names(expr: &Expr, written: &mut Vec<String>) {
-    match &expr.kind {
-        ExprKind::PreIncrement(name)
-        | ExprKind::PostIncrement(name)
-        | ExprKind::PreDecrement(name)
-        | ExprKind::PostDecrement(name) => push_written_name(written, name),
-        _ => {}
-    }
-}
-
-fn push_written_name(written: &mut Vec<String>, name: &str) {
-    if !written.iter().any(|known| known == name) {
-        written.push(name.to_string());
     }
 }
